adde mocks to resolver tests

+++ src/storefront/entities/storefront.entity.ts
@@ -34,6 +34,6 @@ export class Storefront {
   @Column('int', { array: true, nullable: false })
   zip: number[];
   @Field(() => [Int])
-  @Column('bigint', { array: true, nullable: false })
+  @Column('int', { array: true, nullable: false })
   coupons: number[];
 }
